backend/ocr/requirements.txt
fastapi
uvicorn[standard]
pillow
paddleocr
paddlepaddle
python-multipart